apps/api/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
django-filter==23.5
dj-database-url==2.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
boto3==1.34.0
botocore==1.34.0
Pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.220
easyocr==1.7.0
drf-spectacular==0.27.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0